Add webpack loader for external CSS files (#466)
In v3 we improved how external styles work and as a result of that now with a loader like this we support external .css files. 

This means that one can write your styles in a regular .css file to benefit from all the existing tooling (including preprocessors, autocompletion, syntax highlighting) and then consume them with styled-jsx.